# openapi.yaml
openapi: 3.1.0
info:
  title: MCP TypeScript Simple API
  version: 1.0.0
  description: |
    Production-ready Model Context Protocol (MCP) server with OAuth 2.0 authentication,
    Dynamic Client Registration (RFC 7591), and multi-LLM integration.

    ## Features
    - **Dual-mode operation**: STDIO (traditional) + Streamable HTTP with OAuth
    - **Multi-LLM integration**: Claude, OpenAI, and Gemini with type-safe provider selection
    - **OAuth Dynamic Client Registration (DCR)**: RFC 7591 compliant automatic client registration
    - **Comprehensive discovery**: RFC 8414, RFC 9728, and OpenID Connect metadata
    - **OpenTelemetry observability**: Structured logging, metrics, and tracing

    ## Authentication
    This API uses OAuth 2.0 for authentication. Use the `/auth/{provider}/authorize` endpoint
    to initiate the OAuth flow, then use the returned access token with Bearer authentication
    for subsequent requests.

    ## MCP Protocol
    The main MCP endpoint (`/mcp`) implements JSON-RPC 2.0 over Streamable HTTP transport.
    All tool invocations use the JSON-RPC format with `method: tools/call`.

  contact:
    name: Jeff Dutton
    url: https://github.com/jdutton/mcp-typescript-simple
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://mcp-typescript-simple.vercel.app
    description: Production server (Vercel)
  - url: https://{preview-id}.vercel.app
    description: Preview deployment (Vercel)
    variables:
...
security:
  - BearerAuth: []
  - OAuth2: []
paths:
...
  /admin/sessions:
    get:
      tags:
        - Admin & Monitoring
      summary: List active sessions
      description: Returns list of active MCP sessions
      operationId: listSessions
      responses:
        '200':
          description: Active sessions list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionList'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/sessions/{sessionId}:
    delete:
      tags:
        - Admin & Monitoring
      summary: Close specific session
      description: Forcefully closes a specific MCP session
      operationId: closeSession
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session closed successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/info:
    get:
      tags:
        - Admin & Monitoring
      summary: Get deployment information
      description: Returns deployment configuration and environment details
      operationId: getDeploymentInfo
      responses:
        '200':
          description: Deployment information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token authentication using OAuth 2.0 access tokens

    OAuth2:
      type: oauth2
      description: OAuth 2.0 authorization code flow with PKCE
      flows:
        authorizationCode:
          authorizationUrl: /auth/google/authorize
          tokenUrl: /token
          refreshUrl: /token
          scopes:
            openid: OpenID Connect scope
            profile: User profile information
            email: User email address

    ClientAuth:
      type: http
      scheme: bearer
      description: Client authentication using client_secret as bearer token
  schemas:
...
    SessionList:
      type: object
      properties:
        sessions:
          type: array
          items:
            type: object
            properties:
              sessionId:
                type: string
                format: uuid
              authStatus:
                type: string
                enum: [anonymous, authenticated]
              createdAt:
                type: string
                format: date-time
              lastActivity:
                type: string
                format: date-time
        total:
          type: integer
...
    DeploymentInfo:
      type: object
      properties:
        platform:
          type: string
          description: Deployment platform (e.g., 'vercel', 'local')
          example: vercel
        mode:
          type: string
          description: Deployment mode (e.g., 'serverless', 'local')
          example: serverless
        version:
          type: string
          description: Application version
          example: 1.0.0
        node_version:
          type: string
          description: Node.js version
          example: v20
        region:
          type: string
          description: Deployment region
          example: iad1
        deployment_id:
          type: string
          description: Unique deployment identifier
          example: dpl_abc123
        deployment_url:
          type: string
          description: Deployment URL
          example: https://mcp-typescript-simple.vercel.app
        git_commit:
          type: string
          description: Git commit SHA
          example: 7f7dab7
        git_branch:
          type: string
          description: Git branch name
          example: main
        environment:
          type: string
          description: Environment (development, production, etc.)
          example: production
        uptime:
          type: number
          description: Server uptime in seconds
          example: 123.45
        memory_usage:
          type: object
          description: Memory usage statistics
          properties:
            rss:
              type: integer
            heapTotal:
              type: integer
            heapUsed:
              type: integer
            external:
              type: integer
            arrayBuffers:
              type: integer
        cpu_usage:
          type: object
          description: CPU usage statistics
          properties:
            user:
              type: integer
            system:
              type: integer
        oauth_providers:
          type: array
          description: Configured OAuth providers
          items:
            type: string
          example: ['google', 'github']
        oauth_configured:
          type: boolean
          description: Whether OAuth is configured
          example: true
        llm_providers:
          type: array
          description: Configured LLM providers
          items:
            type: string
          example: ['claude', 'openai']
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...
  responses:
...
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingToken:
              value:
                error: "unauthorized"
                message: "Missing or invalid access token"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            notFound:
              value:
                error: "not_found"
                message: "Requested resource not found"
...